합치기 전
