+++ src/component/game/NormalRoom.css
@@ -114,6 +114,7 @@
     background-color: #D9D9D9;
     border-radius: 15px;
     cursor: pointer;
+    color: black;
 }
 
 .joinButton:hover {
