+++ src/component/game/PrivateRoom.css
@@ -132,6 +132,7 @@ input:focus {
     background-color: #D9D9D9;
     border-radius: 15px;
     cursor: pointer;
+    color: black;
 }
 .privateJoinButton:hover {
     background-color: #9F9F9F;
